requirements.txt
requests==2.31.0
beautifulsoup4==4.12.2
python-telegram-bot==20.6
asyncio